+++ misc/memory-connection-limits/Cargo.toml
@@ -1,18 +1,21 @@
 [package]
-name = "libp2p-connection-limits"
+name = "libp2p-memory-connection-limits"
 edition = "2021"
 rust-version = { workspace = true }
-description = "Connection limits for libp2p."
-version = "0.2.1"
+description = "Memory usage based connection limits for libp2p."
+version = "0.1.0"
 license = "MIT"
 repository = "https://github.com/libp2p/rust-libp2p"
 keywords = ["peer-to-peer", "libp2p", "networking"]
 categories = ["network-programming", "asynchronous"]
 
 [dependencies]
+memory-stats = "1"
 libp2p-core = { workspace = true }
 libp2p-swarm = { workspace = true }
 libp2p-identity = { workspace = true, features = ["peerid"] }
+log = "0.4"
+sysinfo = "0.29"
 void = "1"
 
 [dev-dependencies]
